Refactored package name Added MusicConfig Added META-INF to gitignore
