+++ .gitignore
@@ -30,3 +30,5 @@ config/
 commandresources/
 test/
 /TODO
+/src/main/resources/META-INF/MANIFEST.MF
+/src/main/java/META-INF/MANIFEST.MF
