--- src/main/java/com/raiku/troubleclub/radio/commands/PlayTop.java
@@ -1,104 +1,0 @@
-package com.raiku.troubleclub.radio.commands;
-
-import com.raiku.troubleclub.radio.music.PlayerManager;
-import com.raikuman.botutilities.commands.manager.CommandContext;
-import com.raikuman.botutilities.commands.manager.CommandInterface;
-import com.raikuman.botutilities.helpers.MessageResources;
-import net.dv8tion.jda.api.entities.GuildVoiceState;
-import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.api.entities.TextChannel;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-
-/**
- * Handles adding a new song to the top of the queue
- *
- * @version 1.1 2020-23-06
- * @since 1.0
- */
-public class PlayTop implements CommandInterface {
-
-	@Override
-	public void handle(CommandContext ctx) {
-		final TextChannel channel = ctx.getChannel();
-		final Member self = ctx.getGuild().getSelfMember();
-		final GuildVoiceState selfVoiceState = self.getVoiceState();
-		if (selfVoiceState == null) {
-			MessageResources.connectError(channel, 5);
-			return;
-		}
-
-		GuildVoiceState memberVoiceState = ctx.getEventMember().getVoiceState();
-		if (memberVoiceState == null) {
-			MessageResources.connectError(channel, 5);
-			return;
-		}
-
-		if (!memberVoiceState.inAudioChannel()) {
-			MessageResources.timedMessage(
-				"You must be in a voice channel to use this command",
-				channel,
-				5
-			);
-			return;
-		}
-
-		if (selfVoiceState.inAudioChannel() && (selfVoiceState.getChannel() != memberVoiceState.getChannel())) {
-			if (selfVoiceState.getChannel() == null) {
-				MessageResources.connectError(channel, 5);
-				return;
-			}
-
-			MessageResources.timedMessage(
-				"You must be in `" + selfVoiceState.getChannel().getName() + "` to use this command",
-				channel,
-				5
-			);
-			return;
-		}
-
-		String link = String.join(" ", ctx.getArgs());
-		if (!isUrl(link))
-			link = "ytsearch:" + link;
-
-		PlayerManager.getInstance().loadToTop(channel, link, ctx.getEvent().getAuthor(), false);
-
-		ctx.getEvent().getMessage().delete().queue();
-	}
-
-	@Override
-	public String getInvoke() {
-		return "playtop";
-	}
-
-	@Override
-	public String getUsage() {
-		return "<link>";
-	}
-
-	@Override
-	public String getDescription() {
-		return "Adds a song from a link or playlist, or a searched song to the top of the queue";
-	}
-
-	@Override
-	public List<String> getAliases() {
-		return List.of("pt");
-	}
-
-	/**
-	 * Check if the string input is a url
-	 * @param url Url string
-	 * @return If string is url
-	 */
-	private boolean isUrl(String url) {
-		try {
-			new URI(url);
-			return true;
-		} catch (URISyntaxException e) {
-			return false;
-		}
-	}
-}
