--- src/main/java/com/raiku/troubleclub/radio/commands/Shuffle.java
@@ -1,94 +1,0 @@
-package com.raiku.troubleclub.radio.commands;
-
-import com.raiku.troubleclub.radio.music.GuildMusicManager;
-import com.raiku.troubleclub.radio.music.PlayerManager;
-import com.raikuman.botutilities.commands.manager.CommandContext;
-import com.raikuman.botutilities.commands.manager.CommandInterface;
-import com.raikuman.botutilities.helpers.MessageResources;
-import com.raikuman.botutilities.helpers.RandomColor;
-import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.entities.GuildVoiceState;
-import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.api.entities.TextChannel;
-
-public class Shuffle implements CommandInterface {
-
-	@Override
-	public void handle(CommandContext ctx) {
-		final TextChannel channel = ctx.getChannel();
-		final Member self = ctx.getGuild().getSelfMember();
-		final GuildVoiceState selfVoiceState = self.getVoiceState();
-		if (selfVoiceState == null) {
-			MessageResources.connectError(channel, 5);
-			return;
-		}
-
-		GuildVoiceState memberVoiceState = ctx.getEventMember().getVoiceState();
-		if (memberVoiceState == null) {
-			MessageResources.connectError(channel, 5);
-			return;
-		}
-
-		if (!memberVoiceState.inAudioChannel()) {
-			MessageResources.timedMessage(
-				"You must be in a voice channel to use this command",
-				channel,
-				5
-			);
-			return;
-		}
-
-		if (selfVoiceState.inAudioChannel() && (selfVoiceState.getChannel() != memberVoiceState.getChannel())) {
-			if (selfVoiceState.getChannel() == null) {
-				MessageResources.connectError(channel, 5);
-				return;
-			}
-
-			MessageResources.timedMessage(
-				"You must be in `" + selfVoiceState.getChannel().getName() + "` to use this command",
-				channel,
-				5
-			);
-			return;
-		}
-
-		final GuildMusicManager musicManager = PlayerManager.getInstance().getMusicManager(ctx.getGuild());
-
-		if (musicManager.trackScheduler.queue.size() == 0) {
-			MessageResources.timedMessage(
-				"There's currently nothing in the queue",
-				channel,
-				5
-			);
-			return;
-		}
-
-		int songsShuffled = musicManager.trackScheduler.shuffle();
-
-		EmbedBuilder builder = new EmbedBuilder()
-			.setColor(RandomColor.getRandomColor())
-			.setAuthor("\uD83D\uDD00 Shuffled " + songsShuffled + " songs from the queue",
-				null,
-				ctx.getEventMember().getEffectiveAvatarUrl()
-			);
-
-		ctx.getChannel().sendMessageEmbeds(builder.build()).queue();
-
-		ctx.getEvent().getMessage().delete().queue();
-	}
-
-	@Override
-	public String getInvoke() {
-		return "shuffle";
-	}
-
-	@Override
-	public String getUsage() {
-		return "";
-	}
-
-	@Override
-	public String getDescription() {
-		return "Shuffles the current queue";
-	}
-}
