+++ src/main/java/com/raikuman/troubleclub/radio/commands/Prune.java
@@ -1,0 +1,89 @@
+package com.raikuman.troubleclub.radio.commands;
+
+import com.raikuman.troubleclub.radio.music.GuildMusicManager;
+import com.raikuman.troubleclub.radio.music.PlayerManager;
+import com.raikuman.botutilities.commands.manager.CommandContext;
+import com.raikuman.botutilities.commands.manager.CommandInterface;
+import com.raikuman.botutilities.helpers.MessageResources;
+import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.entities.GuildVoiceState;
+import net.dv8tion.jda.api.entities.Member;
+import net.dv8tion.jda.api.entities.TextChannel;
+
+/**
+ * Handles removing duplicate songs from the current queue
+ *
+ * @version 1.1 2020-23-06
+ * @since 1.0
+ */
+public class Prune implements CommandInterface {
+
+	@Override
+	public void handle(CommandContext ctx) {
+		final TextChannel channel = ctx.getChannel();
+		final Member self = ctx.getGuild().getSelfMember();
+		final GuildVoiceState selfVoiceState = self.getVoiceState();
+		if (selfVoiceState == null) {
+			MessageResources.connectError(channel, 5);
+			return;
+		}
+
+		GuildVoiceState memberVoiceState = ctx.getEventMember().getVoiceState();
+		if (memberVoiceState == null) {
+			MessageResources.connectError(channel, 5);
+			return;
+		}
+
+		if (!memberVoiceState.inAudioChannel()) {
+			MessageResources.timedMessage(
+				"You must be in a voice channel to use this command",
+				channel,
+				5
+			);
+			return;
+		}
+
+		if (selfVoiceState.inAudioChannel() && (selfVoiceState.getChannel() != memberVoiceState.getChannel())) {
+			if (selfVoiceState.getChannel() == null) {
+				MessageResources.connectError(channel, 5);
+				return;
+			}
+
+			MessageResources.timedMessage(
+				"You must be in `" + selfVoiceState.getChannel().getName() + "` to use this command",
+				channel,
+				5
+			);
+			return;
+		}
+
+		final GuildMusicManager musicManager = PlayerManager.getInstance().getMusicManager(ctx.getGuild());
+
+		int songsPruned = musicManager.trackScheduler.pruneTracks();
+
+		EmbedBuilder builder = new EmbedBuilder()
+			.setAuthor("✂️ Pruned " + songsPruned + " songs from the queue",
+				null,
+				ctx.getEventMember().getEffectiveAvatarUrl()
+			);
+
+		ctx.getChannel().sendMessageEmbeds(builder.build()).queue();
+
+		ctx.getEvent().getMessage().delete().queue();
+	}
+
+	@Override
+	public String getInvoke() {
+		return "prune";
+	}
+
+	@Override
+	public String getUsage() {
+		return "";
+	}
+
+	@Override
+	public String getDescription() {
+		return "Removes duplicate songs from the queue";
+	}
+}
