+++ src/main/java/com/raikuman/troubleclub/radio/music/AudioPlayerSendHandler.java
@@ -1,0 +1,42 @@
+package com.raikuman.troubleclub.radio.music;
+
+import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
+import com.sedmelluq.discord.lavaplayer.track.playback.MutableAudioFrame;
+import net.dv8tion.jda.api.audio.AudioSendHandler;
+
+import java.nio.ByteBuffer;
+
+/**
+ * Provides information on the audio send handler to provide audio
+ *
+ * @version 1.0 2022-23-06
+ * @since 1.0
+ */
+public class AudioPlayerSendHandler implements AudioSendHandler {
+
+	private final AudioPlayer audioPlayer;
+	private final ByteBuffer buffer;
+	private final MutableAudioFrame frame;
+
+	public AudioPlayerSendHandler(AudioPlayer audioPlayer) {
+		this.audioPlayer = audioPlayer;
+		this.buffer = ByteBuffer.allocate(1024);
+		this.frame = new MutableAudioFrame();
+		this.frame.setBuffer(buffer);
+	}
+
+	@Override
+	public boolean canProvide() {
+		return this.audioPlayer.provide(this.frame);
+	}
+
+	@Override
+	public ByteBuffer provide20MsAudio() {
+		return this.buffer.flip();
+	}
+
+	@Override
+	public boolean isOpus() {
+		return true;
+	}
+}
